Makefile.PL: help to find linking libraries on Gentoo
On Gentoo we install oracle-instantclient libraries in
/usr/lib64/oracle/client/lib64 on 64-bit systems and in
/usr/lib/oracle/client/lib on 32-bit systems.

This was done because in a 64-bit system with multilib,
both 32-bit libraries (/usr/lib64/oracle/client/lib) and
64-bit libraries (/usr/lib64/oracle/client/lib64) were
installed. For now, this specification is disabled because
it is rare to have the same version of oracle instantclient
for both 32-bit and 64-bit.